refactor: update navigation types
