+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/TabNav_TicketingStack.tsx
@@ -3,19 +3,8 @@ import {createStackNavigator} from '@react-navigation/stack';
 import {Ticketing_NotEnabledScreen} from './Ticketing_NotEnabledScreen';
 import {Ticketing_RootScreen} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_RootScreen';
 import {useRemoteConfig} from '@atb/RemoteConfigContext';
-import {NavigatorScreenParams} from '@react-navigation/native';
-import {
-  TicketTabNavStackParams,
-  TicketingTicketHistoryScreenParams,
-} from './Ticketing_TicketTabNavStack/navigation-types';
-import {StackParams} from '@atb/stacks-hierarchy/navigation-types';
 import {Ticketing_TicketHistoryScreen} from './Ticketing_TicketHistoryScreen';
-
-export type TicketingStackParams = StackParams<{
-  Ticketing_NotEnabledScreen: undefined;
-  Ticketing_RootScreen: NavigatorScreenParams<TicketTabNavStackParams>;
-  Ticketing_TicketHistoryScreen: TicketingTicketHistoryScreenParams;
-}>;
+import {TicketingStackParams} from './navigation-types';
 
 const Stack = createStackNavigator<TicketingStackParams>();
