+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/navigation-types.ts
@@ -1,23 +1,22 @@
 import {CompositeScreenProps} from '@react-navigation/native';
 import {StackScreenProps} from '@react-navigation/stack';
-import {TabNavigatorScreenProps} from '@atb/stacks-hierarchy/Root_TabNavigatorStack';
 import {StackParams} from '@atb/stacks-hierarchy/navigation-types';
+import {TicketingScreenProps} from '../navigation-types';
 
 export type TicketTabNavStackParams = StackParams<{
   TicketTabNav_PurchaseTabScreen: undefined;
   TicketTabNav_ActiveFareProductsTabScreen: undefined;
-  Ticketing_TicketHistoryScreen: TicketingTicketHistoryScreenParams;
 }>;
 
 export type TicketingTicketHistoryScreenParams = {
   mode: 'expired' | 'sent';
 };
 
-export type TicketingStackRootProps =
-  TabNavigatorScreenProps<'TabNav_TicketingStack'>;
+export type TicketTabNavStackRootProps =
+  TicketingScreenProps<'Ticketing_RootScreen'>;
 
 export type TicketTabNavScreenProps<T extends keyof TicketTabNavStackParams> =
   CompositeScreenProps<
     StackScreenProps<TicketTabNavStackParams, T>,
-    TicketingStackRootProps
+    TicketTabNavStackRootProps
   >;
